Add basic integration with ML-DSA and EVP
This just makes the single-shot signing APIs work. There's no support in
EVP yet for external mu, context signing, or init/update/final. They are
also not (yet) wired up to the default parsers or EVP_PKEY_CTX_new_id.
You have to use EVP_PKEY_ALG for that.

We only support the seed-based syntax for private keys.

This required adding a CheckVerify codepath to the SignMessage branch
of evp_test. This is very repetitive due to problems with the OpenSSL
API. They choose to make SignMessage use a completely different set of
types from SignDigest. Our original API proposal
(https://boringssl-review.googlesource.com/c/boringssl/+/14447) did not
have this problem, but then OpenSSL requested it be changed to this
form.

Amusingly, OpenSSL seems to have since changed their mind and added
something that looks like our original API. We'll probably need to
restore that scheme to add init/update/final hooks for ML-DSA. See
go/mldsa-mlkem-evp

I've also not yet added keygen support, not because it's particularly
difficult, but because we don't have hooks yet for testing it, or a
clear answer for the static-linker-friendly API. That's probably the
next thing to fill in.

Bug: 449751916
Change-Id: I02d2b5924983a0b9c7f9e71e96de17db687c2425
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/82992
Reviewed-by: Adam Langley <agl@google.com>
Commit-Queue: David Benjamin <davidben@google.com>