Deprecate no-op startup/shutdown ERR functions
Functions like this are usually in the deprecated section. This was
probably early enough that we hadn't actually set this convention.

(This is deprecated in the sense of "don't call this, it's pointless",
not "we actively plan to remove it". To that end, remove the TODO to
remove ERR_load_BIO_strings. Quite a lot more than libjingle calls it,
so I doubt we will ever do away with it.)

Change-Id: Ia13fede6a4b7e25ae37072866acf773c5098989c
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/84387
Reviewed-by: Lily Chen <chlily@google.com>
Commit-Queue: Lily Chen <chlily@google.com>
Auto-Submit: David Benjamin <davidben@google.com>