Add a utility to extract identifiers from public headers.
This is experimental for now, and will later be used or changed to help
with symbol prefixing. The ultimate goal is finding a way to safely link
BoringSSL and OpenSSL into the same binary, or even two BoringSSL
versions; this tool is part of the effort of first analyzing the public
symbols BoringSSL is providing to figure out next steps.

Tested:
  ```
  for f in include/openssl/*.h; do echo "#include <${f#include/}>"; done |\
    clang++ -x c++ -std=c++17 -Iinclude -fsyntax-only -Xclang -ast-dump=json - \
    go run util/extract_identifiers_clang_json.go > extract_identifiers.txt
  ```
  has identical output and no errors on both Linux and MacOS from
  clang-16 to clang-next (which will become clang-22).

Bug: 42220000
Change-Id: I3b3228242b29b8d97fa3a059e8e24caf91f39901
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/84708
Commit-Queue: Adam Langley <agl@google.com>
Reviewed-by: Adam Langley <agl@google.com>
Auto-Submit: Rudolf Polzer <rpolzer@google.com>