infra: Prefix all of BoringSSL's caches
It's unclear if this is the root cause of the mac_arm64_bazel problems
or just a trigger, but the swarming.cache docs suggest prefixing all our
cache names if we share pools with other projects, and we do:
https://chromium.googlesource.com/infra/luci/luci-go/+/refs/heads/main/lucicfg/doc/README.md#swarming.cache

It's possible some of these caches can actually be safely shared.
I'm not aware of any caching issues with $GOCACHE for instance. But it
is still probably better to isolate this than not.

Bug: 463446310, 466089105
Change-Id: If2d8eef885755a01cc9083eab0da7373389ed8e7
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/85428
Auto-Submit: David Benjamin <davidben@google.com>
Commit-Queue: Lily Chen <chlily@google.com>
Reviewed-by: Lily Chen <chlily@google.com>