infra: Remove seemingly unused win_toolchain cache
I can't find any reference to a win_toolchain directory in our repo or
the recipe. This seems to date to when we first moved to lucicfg, and
how we've set up the Windows toolchain has changed a few times, so this
is probably just a remnant. (We currently get the Windows toolchain out
of CIPD.)

Change-Id: Iaaa3c14314a3359a3e638c3af87f5664d401819a
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/85429
Commit-Queue: Lily Chen <chlily@google.com>
Reviewed-by: Lily Chen <chlily@google.com>
Auto-Submit: David Benjamin <davidben@google.com>