Recreate the TestIOVecs on every test call
The in-place tests currently do not work right. Since the output aliases
the input, once we've run the test once, the input has been corrupted
and the tag is invalid anyway.

This was only working on accident because it happened that all the reuse
cases were cases with corrupted tags. But the tests were not actually
testing a corrupted tag, because it wasn't the riht tag to begin with.

Bug: 383343306
Change-Id: I43f58eb15ed0ba9b654499795e896ea6d09b466c
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/85507
Commit-Queue: Rudolf Polzer <rpolzer@google.com>
Auto-Submit: David Benjamin <davidben@google.com>
Reviewed-by: Rudolf Polzer <rpolzer@google.com>