Split sealv, openv, and openv_detached into separate tests
Our tests currently run sealv+openv and sealv+openv_detached as two
separate tests, but this means the sealv portion is redundant between
the two and run twice.

It's also a bit confusing because we put a lot of effort into using the
sealv'd output, but we've already tested that it's identical to the test
vector. Splitting it makes it clearer we're testing two separate
operations, and we don't run the sealv multiplicative tests twice.

Bug: 383343306
Change-Id: Iff8cddafb91dd118be92e30e7bd88572930faad1
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/85508
Commit-Queue: Rudolf Polzer <rpolzer@google.com>
Reviewed-by: Rudolf Polzer <rpolzer@google.com>
Auto-Submit: David Benjamin <davidben@google.com>