Add a script to list unintended symbols in libcrypto.
A symbol is for now defined as unintended if it's not declared in a
public header file, not namespaced, and likely seen by the linker.

This will be used as a basis of symbol/namespacing cleanup; however
it may be the case that not all symbols can be fixed easily.

Bug: 42220000
Change-Id: Ic8daebeaba1223110b0a7c20b65bceb79890558d
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/85531
Commit-Queue: Rudolf Polzer <rpolzer@google.com>
Auto-Submit: Rudolf Polzer <rpolzer@google.com>
Reviewed-by: David Benjamin <davidben@google.com>