Fix up constant-time annotations for TLS CBC ciphers
One of the CONSTTIME_SECRET annotations was reudndant. It was already
marked secret right after the call to EVP_DecryptUpdate_ex.

The CONSTTIME_SECRETs at the end should be CONSTTIME_DECLASSIFY because
we're allowing the caller to see the result.

One of the asserts should be declassify_assert because we're going to,
in debug mode, branch on the result.

Change-Id: I9dcbe7c223acc9bbfe13695570e7ada34b82f5d5
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/85627
Auto-Submit: David Benjamin <davidben@google.com>
Commit-Queue: Rudolf Polzer <rpolzer@google.com>
Reviewed-by: Rudolf Polzer <rpolzer@google.com>