Add API to configure server's NamedGroups with equal preference
This adds an API to associate each named group with a "flags" value, and
adds a flag that indicates that the group has equal preference with the
next group in the list. This is analogous to "in_group" flags used with
SSL ciphers (but is renamed to avoid using the overloaded term "group").
When SSL_OP_CIPHER_SERVER_PREFERENCE is specified, the server's
supported group list is used, so these "with next" flags allow the
server to express equal preference between consecutively listed groups,
so that a choice among them respects the client's preference.

Bug: 437414714
Change-Id: I73ec6523d6a6e3d852a5664d49cf1f0a6a6a6964
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/85787
Commit-Queue: David Benjamin <davidben@google.com>
Auto-Submit: Lily Chen <chlily@google.com>
Reviewed-by: David Benjamin <davidben@google.com>