Compute the FIPS module hash after evaluating relocations
Our build currently requires the BCM text segment to have *no*
relocations, but that's a little strong. It is sufficient for it to have
only link-independent relocations. I.e. relocations that always give the
same result no matter what it is linked against.

In a sufficiently smart toolchain, these would be the same, but
sometimes we get these. On AArch64, ADRP always emits a relocation in
Clang's assembler, even when the page offsets are known due to alignment
requirements. The linker script idea in crbug.com/362530616 also emits a
ton of these because `ld -r` does not know how to partially evaluate
relocations.

As part of this, build bcm_hashunset as a CMake object library instead
of a CMake static library. If it's a static library, CMake gets very
upset about the shared library having no sources, and keeps trying to
drop the static library as unused.

Update-Note: See also cl/846024477

Bug: 362530616, 469788763
Change-Id: I2f1999679b1987664da4350d86ac149e52d44303
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/85928
Reviewed-by: Adam Langley <agl@google.com>
Commit-Queue: David Benjamin <davidben@google.com>