Make libssl no longer export the lhash symbols for SSL_SESSION.
This requires making it possible to name LHASH_OF of a type outside the
current namespace; the simplest solution was to use a suffix, not a
prefix, to identify lhash types.

After this, libssl is 100% clean of unnamespaced internal symbols that
don't originate from the other libraries.

Bug: 42220000
Change-Id: Id4fd04b702cf9893b6afd6a4f6311b0469c6aa2b
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/85967
Auto-Submit: Rudolf Polzer <rpolzer@google.com>
Reviewed-by: Adam Langley <agl@google.com>
Commit-Queue: Rudolf Polzer <rpolzer@google.com>