Convert Go string-empty checks to `!= ""`.
Way back, `len(x) != 0` was more efficient than `!= ""` and a lot of the
Go code in BoringSSL is still based on old habits developed in that era.

That's no longer true, so use the (now) more standard form for checking
whether a string is non-empty.

Change-Id: I2e4698e414ed86e68261e7b4c8446fc2603dd6d7
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/86007
Commit-Queue: Adam Langley <agl@google.com>
Reviewed-by: David Benjamin <davidben@google.com>